Added Connect 4 Blogpost
Added details about connect 4 blogpost. Not sure what the goals are going to be for the program. I do want it to be fully functioning game of 'table top' pass and play style of game. Should be easy enough. Could get diffucult if trying to develop an AI or an Online Mode.